examples/memory_example/requirements.txt
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.5.0

# Optional dependencies
python-Levenshtein>=0.12.0  # Improves performance of fuzzywuzzy
